Fix directory structure of math
