+++ pooltool/math/__init__.py
@@ -1,5 +1,3 @@
-import pooltool.math.roots.quadratic as quadratic
-import pooltool.math.roots.quartic as quartic
 from pooltool.math._math import (
     angle,
     coordinate_rotation,
@@ -9,4 +7,3 @@ from pooltool.math._math import (
     unit_vector_slow,
     wiggle,
 )
-from pooltool.math.roots import min_real_root, minimum_quartic_root
